Add solidity formatter

+++ contracts/mocks/TendermintMock.sol
@@ -3,10 +3,9 @@
 pragma solidity ^0.8.2;
 
 import "../proto/TendermintHelper.sol";
-import { SignedHeader, ValidatorSet } from "../proto/TendermintLight.sol";
+import {SignedHeader, ValidatorSet} from "../proto/TendermintLight.sol";
 
 contract TendermintMock {
-
     function signedHeaderHash(bytes memory data) public pure returns (bytes32) {
         SignedHeader.Data memory sh = SignedHeader.decode(data);
         return TendermintHelper.hash(sh);
@@ -17,7 +16,7 @@ contract TendermintMock {
         return TendermintHelper.getTotalVotingPower(vs);
     }
 
-    function getByAddress(bytes memory data, bytes memory addr) public pure returns (uint index, bool found) {
+    function getByAddress(bytes memory data, bytes memory addr) public pure returns (uint256 index, bool found) {
         ValidatorSet.Data memory vs = ValidatorSet.decode(data);
         return TendermintHelper.getByAddress(vs, addr);
     }
